requirements.txt
Django==4.2.3
djangorestframework==3.14
Faker==v19.6.1
Pillow==10.0.1
ruff